Fragment multipoles writes to file now.
This involved modifyign bigdft-tool and utitiles to support that option,
however it is an optional parameters.

It was necessary to add this feature because file redirection of mpi
executables does not work on all systems (the K computer...). In general,
it seems like the main drivers of BigDFT should be able to write to
a log like file.